Refactor the transect table that shows up on import, add row coloring
