+++ src/main/python/importwizards/transecttable/__init__.py
@@ -1,0 +1,4 @@
+
+from .transect import Transect
+from .transectmodel import TransectTableModel
+from .transectview import TransectTableView
