+++ src/main/python/transects/__init__.py
@@ -1,4 +1,2 @@
 
-from .transectmodel import Transect, TransectTableModel
-from .transectview import TransectTableView
 from .saves import TransectSaveData, TransectSaveDatas
